requirements.txt
discord.py==1.6.0
peewee==3.14.4